# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/master": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "pod name",
                        "name": "key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/master/copy": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "CopyRequest",
                        "name": "podBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CopyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/master/getconfig": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/master/getenv": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/master/partition": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "PartitionRequest",
                        "name": "podBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PartitionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/master/search": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "SearchRequest",
                        "name": "podBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CopyRequest": {
            "type": "object",
            "properties": {
                "fromId": {
                    "type": "integer"
                },
                "fromTopic": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "podName": {
                    "type": "string"
                },
                "toId": {
                    "type": "integer"
                },
                "toTopic": {
                    "type": "string"
                }
            }
        },
        "dto.PartitionRequest": {
            "type": "object",
            "properties": {
                "clusterId": {
                    "type": "integer"
                },
                "groupId": {
                    "type": "string"
                },
                "ignoredPartitions": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "key": {
                    "type": "string"
                },
                "podName": {
                    "type": "string"
                },
                "topic": {
                    "type": "string"
                }
            }
        },
        "dto.SearchRequest": {
            "type": "object",
            "properties": {
                "endDate": {
                    "type": "integer"
                },
                "kafkaId": {
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "metadataKey": {
                    "type": "string"
                },
                "podName": {
                    "type": "string"
                },
                "startDate": {
                    "type": "integer"
                },
                "topic": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                },
                "valueType": {
                    "type": "integer"
                }
            }
        }
    }
}